+++ src/tr1/game/output.c
@@ -386,9 +386,10 @@ void Output_SetWindowSize(int32_t width, int32_t height)
 
 void Output_ApplyLevelSettings(void)
 {
+    const RGB_888 color = Level_GetWaterColor();
+    Output_SetWaterColor(Level_GetWaterColor());
     Output_SetDrawDistFade(Level_GetFogStart() * WALL_L);
     Output_SetDrawDistMax(Level_GetFogEnd() * WALL_L);
-    LOG_INFO("%f %f", Level_GetFogStart(), Level_GetFogEnd());
 }
 
 void Output_ApplyRenderSettings(void)